Fixed some warnings and display text

+++ SFML Projects/SFML_Basic_Concepts/include/LevelLoader.h
@@ -15,7 +15,7 @@ struct ObstacleData
 {
 	std::string m_type;
 	sf::Vector2f m_position;
-	double m_rotation;
+	float m_rotation;
 };
 
 /// <summary>
@@ -28,7 +28,7 @@ struct TargetData
 	sf::Vector2f m_position;
 	int m_randomOffset;
 	float m_durationSeconds;
-	double m_rotation;
+	float m_rotation;
 };
 
 /// <summary>
